gardnx_backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
Pillow==11.0.0
firebase-admin==6.6.0
httpx==0.28.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
pytest==8.3.0
pytest-asyncio==0.24.0
ruff==0.8.0
google-generativeai==0.8.3
